requirements.txt
amqp==2.5.2
apns2==0.5.0
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.229
botocore==1.12.229
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cryptography==2.8
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.4
django-appconf==1.0.3
django-ckeditor==5.7.1
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-excel==0.0.10
django-inline-svg==0.1.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-phone-field==1.8.0
django-push-notifications==1.6.1
django-qr-code==1.0.0
django-rest-auth==0.9.5
django-storages==1.7.2
django-tz-detect==0.2.10
djangorestframework==3.10.2
djongo==1.2.35
docutils==0.15.2
et-xmlfile==1.0.1
# excel2table==2.3.2
ffmpy==0.2.2
fpdf==1.7.2
futures==3.1.1
giftbit==0.1
googlemaps==3.1.3
googletrans==2.4.0
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
img2pdf==0.3.3
importlib-metadata==0.23
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.9.4
jupyter-pip==0.3.1
kombu==4.6.6
lml==0.0.9
lxml==4.5.2
MarkupSafe==1.1.1
mongo==0.2.0
more-itertools==7.2.0
mtranslate==1.6
numpy==1.17.2
odfpy==1.4.1
openpyxl==3.0.1
pilkit==2.0
Pillow==6.1.0
promise==2.2.1
py-vapid==1.7.0
pycparser==2.19
pyecharts-jupyter-installer==0.0.3
pyexcel==0.5.15
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.20
pyexcel-ods==0.5.6
pyexcel-odsr==0.5.2
pyexcel-sortable==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
PyJWT==1.7.1
pymongo==3.9.0
PyMsgBox==1.0.7
PyQRCode==1.2.1
python-dateutil==2.8.0
pytz==2019.2
pywebpush==1.10.0
qrcode==6.1
redis==3.4.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.2.4
texttable==1.6.2
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
